backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
anthropic==0.52.0
PyJWT[crypto]==2.10.1
httpx==0.28.1
pydantic-settings==2.8.1
